Put related things closer to one another
To make project a bit easier to discover it would be helpful if
related things lived close to one another. It also allows to use
more shorter names of scripts and classes. It reduces cognitive load
to separate deepcubes models from services's models and so on.
Here is the list of changes:

* Tests for scripts moved to scripts directory
* NetworkEmbedder should live in scripts as it used only by services
* is_url is used only by NetworkEmbedder, so it moved in the same module
* get_new_model_id is the implementation detail of services, so it moved
  to scripts

Touched filed are formated to follow to PEP-8.